+++ tests/testthat/test-accessors.R
@@ -247,11 +247,32 @@ test_that("quarters accessor extracts correct quarter", {
   expect_equal(quarter(date, with_year = TRUE, fiscal_start = -2), 2011.1)
   expect_equal(quarter(date, with_year = TRUE, fiscal_start = 11), 2011.1)
 
-  x <- ymd(c("2012-03-26", "2012-05-04", "2012-09-23", "2012-03-01", "2012-12-31"))
+  x <- ymd(c("2012-03-01", "2012-03-26", "2012-05-04", "2012-09-23", "2012-12-31"))
   expect_equal(quarter(x, with_year = TRUE, fiscal_start = 4),
-               c(2011.4, 2012.1, 2012.2, 2011.4, 2012.3))
+               c(2012.4, 2012.4, 2013.1, 2013.2, 2013.3))
   expect_equal(quarter(x, with_year = TRUE, fiscal_start = 11),
-               c(2012.2, 2012.3, 2012.4, 2012.2, 2013.1))
+               c(2012.2, 2012.2, 2012.3, 2012.4, 2013.1))
+
+  x <- ymd("2010-01-01") + months(0:11)
+  expect_equal(quarter(x, with_year = TRUE, fiscal_start = 6),
+               c(2010.3, 2010.3, 2010.4, 2010.4, 2010.4, 2011.1, 2011.1, 2011.1,
+                 2011.2, 2011.2, 2011.2, 2011.3))
+  expect_equal(quarter(x, with_year = TRUE, fiscal_start = 6),
+               quarter(x, with_year = TRUE, fiscal_start = -6))
+  x <- ymd("2010-01-01") + months(seq(0, 23, by = 3))
+  expect_equal(quarter(x, with_year = TRUE, fiscal_start = 10),
+               c(2010.2, 2010.3, 2010.4, 2011.1, 2011.2, 2011.3, 2011.4, 2012.1))
+  expect_equal(quarter(x, with_year = TRUE, fiscal_start = -2),
+               quarter(x, with_year = TRUE, fiscal_start = 10))
+
+  x <- ymd(c("2018-07-15", "2018-12-27", "2019-01-01",
+             "2019-06-01", "2019-07-01", "2019-10-16",
+             "2019-12-31", "2020-01-01", "2020-06-30",
+             "2020-07-01", "2020-11-09", "2021-01-19",
+             "2021-06-30", "2021-07-01"))
+  expect_equal(quarter(x, with_year = T, fiscal_start = 7),
+               c(2019.1, 2019.2, 2019.3, 2019.4, 2020.1, 2020.2, 2020.2, 2020.3,
+                 2020.4, 2021.1, 2021.2, 2021.3, 2021.4, 2022.1))
 })
 
 test_that("years accessor extracts correct year", {