+++ NEWS.md
@@ -8,6 +8,7 @@ Version 1.7.8.9000 (dev)
 
 ### BUG FIXES
 
+* [#890](https://github.com/tidyverse/lubridate/issues/890) Correctly compute year in `quarter(..., with_year = TRUE)`
 * [#893](https://github.com/tidyverse/lubridate/issues/893) Fix incorrect parsing of abbreviated months in locales with trailing dot (regression in v1.7.8)
 * [#886](https://github.com/tidyverse/lubridate/issues/886) Fix `with_tz()` for POSIXlt objects
 * [#887](https://github.com/tidyverse/lubridate/issues/887) Error on invalid numeric input to `month()`